rename and move bunch of files
