--- src/lib/types.ts
@@ -1,4 +1,0 @@
-export type FixedLengthArray<T, L extends number> = {
-  0: T;
-  length: L;
-} & ReadonlyArray<T>;
